refactor(contracts): re-structures utils files
