+++ src/contracts/utils/waitForReceipt.ts
@@ -1,6 +1,5 @@
 import Web3 from 'web3'
 import { TransactionReceipt } from 'web3-eth'
-import { SupportedTokens, Token, TOKENS } from './interfaces'
 
 const TIMEOUT_LIMIT = 120000
 const POLLING_INTERVAL = 2000
@@ -28,14 +27,5 @@ function waitForReceipt(
     }, POLLING_INTERVAL)
   })
 }
-
-export const getTokens = (
-  supportedTokens: SupportedTokens[],
-): Token[] => supportedTokens.map((t) => {
-  if (!TOKENS[t]) {
-    throw new Error(`Token ${t} not implemented`)
-  }
-  return TOKENS[t]
-})
 
 export default waitForReceipt
